+++ backend/databasehandler/src/test/java/com/pixtranslator/backend/databasehandler/DictionaryControllerTestSuite.java
@@ -1,0 +1,10 @@
+package com.pixtranslator.backend.databasehandler;
+
+import org.junit.platform.runner.JUnitPlatform;
+import org.junit.platform.suite.api.SelectPackages;
+import org.junit.runner.RunWith;
+
+@RunWith(JUnitPlatform.class)
+@SelectPackages("DictionaryControllerTestPackage")
+public class DictionaryControllerTestSuite {
+}
